homework2/requirements.txt
asgiref==3.7.2
Django==4.2.11
djangorestframework==3.16.1
gunicorn==23.0.0
gyp==0.1
packaging==25.0
pytz==2024.1
setuptools==68.1.2
six==1.16.0
sqlparse==0.4.4
supervisor==4.2.5
wheel==0.42.0
